+++ src/unknown/system.rs
@@ -76,7 +76,7 @@ impl SystemExt for System {
         &[]
     }
 
-    fn get_physical_core_numbers(&self) -> u64 {
+    fn get_physical_core_numbers(&self) -> usize {
         0
     }
